+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ByteArrayUtil.extractBytes` method to extract bytes from a number.
+- `ByteArrayUtil.extractShort` method.
+- `ByteArrayUtil.extractLong` method.
+- `ByteArrayUtil.copyBytes` method.
 
 ## [2.2.0] - 2022-10-03
 ### Added